Module ir_to_bytecode::parser

Functions

is_permitted_char
Determine if a character is permitted character.

is_permitted_newline_char
Determine if a character is a permitted newline character.

is_permitted_printable_char
Determine if a character is an allowed eye-visible (printable) character.

parse_cmd_
Given the raw input of a file, creates a single Cmd_ struct Fails with Err(_) if the text cannot be parsed

parse_module
Given the raw input of a file, creates a single ModuleDefinition struct Fails with Err(_) if the text cannot be parsed

parse_script
Given the raw input of a file, creates a Script struct Fails with Err(_) if the text cannot be parsed

parse_script_or_module
Given the raw input of a file, creates a ScriptOrModule enum Fails with Err(_) if the text cannot be parsed`